requirements.txt
env==0.1.0
mysql-connector-python==8.0.22
protobuf==3.13.0
six==1.15.0
style==1.1.0
update==0.0.1
